add test for stats JSON size
